+++ includes/IPCHandling.php
@@ -1,0 +1,7 @@
+<?php
+  class IPCHandling {
+    public function receiveData($connection, $data) {
+      return;
+    }
+  }
+?>