+++ modules/events/IPCConnectionCreatedEvent.php
@@ -1,0 +1,10 @@
+<?php
+  class __CLASSNAME__ {
+    public $name = "IPCConnectionCreatedEvent";
+
+    public function isInstantiated() {
+      EventHandling::createEvent("IPCConnectionCreatedEvent", $this);
+      return true;
+    }
+  }
+?>