+++ modules/events/IPCConnectionDisconnectedEvent.php
@@ -1,0 +1,10 @@
+<?php
+  class __CLASSNAME__ {
+    public $name = "IPCConnectionDisconnectedEvent";
+
+    public function isInstantiated() {
+      EventHandling::createEvent("IPCConnectionDisconnectedEvent", $this);
+      return true;
+    }
+  }
+?>